remove template queue and added template to regular queue
the template specialization for queue in queue.hpp is currently commented out. this is because for some reason when trying to define it out of line in queue.cpp it gives me an error telling me that no function like that could be found. however, when doing the definition in-line it works. i am currently still trying to figure this one out, if you have any idea how to fix this, please feel free to commit!